Manage dnsmasq api-int record
Add support for updating/removing the api-int record in the dnsmasq
host file. This is necessary so that deployers can point api-int
at their own external loadbalancer if they choose to use one.

Currently this only supports setting the api-int record to the API
VIP or removing it completely and relying on external DNS. In the
future we could support redirecting the API VIP to another address,
but for the initial implementation this is simple and enough to
satisfy the use case.